FE-Refactor: SidePanel 컴포넌트에서 ProfileImageUploader 컴포넌트 분리
